Implemented server for multiple editors to work on one instance of MasterCardAtlas.java Implemented TarotCard model for Card types Updated WildCard and RuneCard models + others (small adjustments) Implemented simple HTTP server for version control of MasterCardAtlas.java Updated form label values to match their variable counterpart for easier reference in the Data Model Fixed typos in forms and default values Updated CharacterBox to CharacterForm Updated CharacterCard model for special abilities Added fool-proof mechanisms in forms for efficient data entry

+++ editor/src/main/java/org/camposmdev/editor/ui/RewardBox.java
@@ -52,10 +52,10 @@ public class RewardBox extends FormController<Reward> {
 
     @Override
     public Reward submit() throws Exception {
-        return new Reward(Byte.parseByte(tfLoot.getText()),
-                Byte.parseByte(tfTreasure.getText()),
-                Byte.parseByte(tfCents.getText()),
-                Byte.parseByte(tfHitPoints.getText()),
+        return new Reward(Integer.parseInt(tfLoot.getText()),
+                Integer.parseInt(tfTreasure.getText()),
+                Integer.parseInt(tfCents.getText()),
+                Integer.parseInt(tfHitPoints.getText()),
                 cbRechargeItem.isSelected(),
                 cbStealShopItem.isSelected(),
                 cbDeath.isSelected(),
